+++ include/Camera.h
@@ -8,13 +8,13 @@
 class Camera : public Object
 {
 public:
-    Camera(size_t width, size_t height, float verticalFieldOfView);
+    Camera(size_t width, size_t height, double verticalFieldOfView);
 
     size_t width() const;
     size_t height() const;
-    float aspectRatio() const;
-    float horizontalFieldOfView() const;
-    float verticalFieldOfView() const;
+    double aspectRatio() const;
+    double horizontalFieldOfView() const;
+    double verticalFieldOfView() const;
 
     std::optional<PixelCoords> coordForPoint(const Vector& point) const;
     Vector pixelDirection(const PixelCoords& coord) const;
@@ -22,7 +22,7 @@ public:
 private:
     const size_t m_width;
     const size_t m_height;
-    const float m_aspectRatio;
-    const float m_verticalFieldOfView;
-    const float m_horizontalFieldOfView;
+    const double m_aspectRatio;
+    const double m_verticalFieldOfView;
+    const double m_horizontalFieldOfView;
 };
