+++ src/AngleGenerator.cpp
@@ -6,26 +6,26 @@
 #include <algorithm>
 #include <cmath>
 
-float AngleGenerator::map(float value) const
+double AngleGenerator::map(double value) const
 {
-    value = std::min(std::max(-1.0f, value), 1.0f);
+    value = std::min(std::max(-1.0, value), 1.0);
     value = std::pow(value, linearity);
     value *= maxAngle;
 
     return value;
 }
 
-float AngleGenerator::generate(RandomGenerator& randomGenerator) const
+double AngleGenerator::generate(RandomGenerator& randomGenerator) const
 {
-    float value = -1.0f + (randomGenerator.value() * 2.0f);
+    double value = -1.0 + (randomGenerator.value() * 2.0);
 
     return map(value);
 }
 
 Vector AngleGenerator::generateOffsetVector(const Vector& center, RandomGenerator& randomGenerator) const
 {
-    float theta = randomGenerator.value(2.0f * Utility::pi);
-    float angle = Utility::radians(map(randomGenerator.value(1.0f)));
+    double theta = randomGenerator.value(2.0 * Utility::pi);
+    double angle = Utility::radians(map(randomGenerator.value(1.0)));
     bool useZ = (center.y > center.x && center.y > center.z);
     const Vector& crossAxis = useZ ? Vector::unitZ : Vector::unitY;
