+++ src/LightQueue.cpp
@@ -3,7 +3,7 @@
 void LightQueue::setPhotonCount(const std::string& name, size_t count)
 {
     std::scoped_lock<std::mutex> lock(m_mutex);
-    m_brightness.insert_or_assign(name, 1.0f / static_cast<float>(count));
+    m_brightness.insert_or_assign(name, 1.0 / static_cast<double>(count));
     m_photons.insert_or_assign(name, count);
     m_remaining.fetch_add(count);
 }
@@ -35,12 +35,12 @@ size_t LightQueue::fetchPhotons(const std::string& name, size_t count)
     return fetched;
 }
 
-float LightQueue::getPhotonBrightness(const std::string& name) const
+double LightQueue::getPhotonBrightness(const std::string& name) const
 {
     std::scoped_lock<std::mutex> lock(m_mutex);
     if (m_brightness.count(name) == 0)
     {
-        return 0.0f;
+        return 0.0;
     }
 
     return m_brightness.at(name);
